+++ src/clj_p4/schema.clj
@@ -236,6 +236,9 @@
    [:lookahead {:optional true} [:maybe [:int {:min 0}]]]
    [:no-merge? {:optional true} [:maybe boolean?]]
    [:keep-empty-commits? {:optional true} [:maybe boolean?]]
+   [:metadata-decoding-strategy {:optional true}
+    [:enum :strict :fallback :passthrough]]
+   [:metadata-fallback-encoding {:optional true} [:maybe string?]]
    [:progress-fn {:optional true} ifn?]
    [:stop? {:optional true} ifn?]])